fix: resolve CI test failures with more robust Obsidian mocking
- Replace virtual mock with alias-based approach for better CI compatibility
- Create dedicated tests/obsidian-stub.ts with comprehensive Obsidian API stubs
- Add missing cachedRead method to vault mock
- Simplify tests/setup.vitest.ts since alias handles module resolution
- Restore obsidian alias in vitest.config.ts pointing to test stub

This approach ensures consistent module resolution across different
environments while maintaining TypeScript type safety in development.
