+++ vitest.config.ts
@@ -13,6 +13,7 @@ export default defineConfig({
 	resolve: {
 		alias: {
 			src: path.resolve("./src"),
+			obsidian: path.resolve("./tests/obsidian-stub.ts"),
 		},
 	},
 	test: {
